Backend/requirements.txt
django>=5.2
djangorestframework
django-cors-headers
openai
google-api-python-client
google-auth
google-auth-httplib2
celery
redis
stripe
